@@ -25,30 +26,24 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-        <spring-boot.version>2.1.9.RELEASE</spring-boot.version>
-        <h2.version>1.4.199</h2.version>
+        <spring-boot.version>2.2.0.RELEASE</spring-boot.version>
         <swagger.version>2.9.2</swagger.version>
-        <jackson.version>2.10.0</jackson.version>
+        <chaos-monkey-spring-boot.version>2.0.2</chaos-monkey-spring-boot.version>
+        <resilience4j.version>1.1.0</resilience4j.version>
+        <junit-jupiter.version>5.5.2</junit-jupiter.version>
+        <junit.platform.version>1.2.0</junit.platform.version>
+        <testcontainers.version>1.12.2</testcontainers.version>
 
         <docker.plugin.version>1.2.0</docker.plugin.version>
         <docker.image.prefix>urthehero</docker.image.prefix>
         <docker.image.exposed.port>9090</docker.image.exposed.port>
-        <docker.image.dockerfile.dir>${basedir}</docker.image.dockerfile.dir>
+        <docker.image.dockerfile.dir>./back</docker.image.dockerfile.dir>
         <docker.image.dockerize.version>v0.6.1</docker.image.dockerize.version>
 
-    </properties>
+        <dockerfile-maven-version>1.4.10</dockerfile-maven-version>
 
-    <dependencies>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.12</version>
-        </dependency>
-    </dependencies>
+        <jacoco.maven.plugin.version>0.8.5</jacoco.maven.plugin.version>
+    </properties>
 
     <dependencyManagement>
         <dependencies>
