@@ -4,6 +4,8 @@ import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import springfox.documentation.RequestHandler;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;
@@ -17,12 +19,15 @@ import java.util.Collections;
 
 @Configuration
 @EnableSwagger2
+@PropertySource(value = {
+        "classpath:/swagger/swagger-messages_fr.properties"
+})
 public class SwaggerConfig {
     @Bean
     public Docket api() {
         return new Docket(DocumentationType.SWAGGER_2)
                 .select()
-                .apis(Predicates.not(RequestHandlerSelectors.basePackage("org.springframework")))
+                .apis(apis())
                 .paths(paths())
                 .build()
                 .apiInfo(apiInfo());
@@ -32,14 +37,17 @@ public class SwaggerConfig {
         return Predicates.not(PathSelectors.regex("/basic-error-controller.*"));
     }
 
+    private Predicate<RequestHandler> apis() {
+        return Predicates.not(RequestHandlerSelectors.basePackage("org.springframework"));
+    }
 
     private ApiInfo apiInfo() {
         return new ApiInfo(
-                "My REST API",
-                "Some custom description of API.",
-                "API TOS",
+                "URtheHero API",
+                "BackEnd API for URtheHero application.",
+                "API 0.0.1-SNAPSHOT",
                 "Terms of service",
-                new Contact("John Doe", "www.example.com", "myeaddress@company.com"),
+                new Contact("Thomas Devile-Duc", "urthehero.tdevilleduc.com", "urthehero.app@gmail.com"),
                 "License of API", "API license URL", Collections.emptyList());
     }
 }
