@@ -1,10 +1,18 @@
 package com.tdevilleduc.urthehero.back.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.*;
 
 import javax.persistence.*;
+import java.util.List;
 
 @Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode
 public class Page {
 
     @Id
@@ -12,52 +20,15 @@ public class Page {
     private Integer id;
 
     @ManyToOne
-    @JoinColumn
+    @JoinColumn(name = "story_Id", insertable = false, updatable = false)
+    @JsonIgnore
     private Story story;
 
     @Column(columnDefinition = "TEXT")
     private String text;
     private String image;
 
-    public Page() {
-    }
+    @Transient
+    private List<NextPage> nextPageList;
 
-    public Page(Integer id, Story story, String text) {
-        this.id = id;
-        this.story = story;
-        this.text = text;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    @JsonIgnore
-    public Story getStory() {
-        return story;
-    }
-
-    public void setStory(Story story) {
-        this.story = story;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
 }
