@@ -1,6 +1,5 @@
 package com.tdevilleduc.urthehero.back.model;
 
 public enum Position {
-
-    LEFT, RIGHT, CENTER;
+    LEFT, RIGHT, CENTER
 }
